requirements.txt
alembic==1.13.2
bidict==0.23.1
blinker==1.8.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.46.0
greenlet==3.0.3
h11==0.14.0
humanfriendly==10.0
idna==3.6
imageio==2.33.0
importlib-resources==6.1.1
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.3
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.4
networkx==3.2.1
numpy==1.23.5
onnxruntime==1.16.3
opencv-python==4.10.0.84
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.25.1
PyMySQL==1.1.1
pyparsing==3.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.31
sympy==1.12
tifffile==2023.9.26
torch==1.8.1
torchvision==0.9.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
zipp==3.17.0